Rename PublicInformation to StopPublicInformation
